Convert to double

+++ Compiler/ControlFlowGraph/IntConstantArgument.cs
@@ -4,12 +4,12 @@
 
     public class IntConstantArgument : Argument
     {
-        public IntConstantArgument(int value)
+        public IntConstantArgument(long value)
         {
             this.Value = value;
         }
 
-        public int Value { get; private set; }
+        public long Value { get; private set; }
 
         public override string ToString()
         {
